1. Added a README file; 2. Made explicit Junit dependency 3. General cleanup

+++ src/main/java/com/nacnez/projects/grid/model/Person.java
@@ -19,7 +19,7 @@ public class Person implements Serializable {
     
     private String gender;
     
-    private BigDecimal income;
+    private Double income;
     
 	private String city;
 
@@ -77,11 +77,11 @@ public class Person implements Serializable {
 		this.address = address;
 	}
 
-	public BigDecimal getIncome() {
+	public Double getIncome() {
 		return income;
 	}
 
-	public void setIncome(BigDecimal income) {
+	public void setIncome(Double income) {
 		this.income = income;
 	}
